Allow entire Tree in memory before persistence
Before this change building up Trees to represent the repository was
extremely difficult due to the "shallow" nature of the TreeRef
struct.  Instead of pointing to a key in the database, this change
allows TreeRefs to represent it's children in-memory, and will only
persist when its parent tree calls .Put().